\name{mapmeta}
\alias{mapmeta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
create a DataFrame instance providing metadata about the
Epigenomics Roadmap
}
\description{
create a DataFrame instance providing metadata about the
Epigenomics Roadmap
}
\usage{
mapmeta()
}
%- maybe also 'usage' for other objects documented here.
\details{
originates at
\url{https://docs.google.com/spreadsheet/ccc?key=0Am6FxqAtrFDwdHU1UC13ZUxKYy1XVEJPUzV6MEtQOXc&usp=sharing#gid=15}
}
\value{
a \code{\link{DataFrame}} instance that is wrapped
to limit sprawl over columns when shown.
}
%\references{
%%% ~put references to the literature/web site here ~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
mapmeta()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }
